requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.2
django-countries==7.5.1
django-crispy-forms==2.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
